simplify code, keep it in sync with the blog post
This introduces a new `addmod` function, that was previously written
inline using the ternary operator. Also, it replaces masking with a
conditional. To my surprise, this performs 24% better for u32, and
20% better for u64 on msvc110, without profile-guided optimisation.
I have yet to try Clang and GCC, but I believe they were faster to
begin with, so I expect no regressions there.