swap argument order again
On my Raspberry Pi, the previous version was better with GCC and
with Clang. However, on my laptop i7 with msvc110, this version is
better. I should still test my desktop i7 with Clang and GCC though,
because previously it turned out that msvc generates far-from-optimal
code, so I might have to reverse it again.